Modify code for sending SOC values through serial port
Change the switch-case statement by if-else statements. It was
introducing a bug in the program that made it unstable, producing
unexpected behaviours which broke the board after the first READY
command.

Set size of output_data accordingly with requested function. Moreover,
the data types of other variables where adapted, to avoid future bugs.