Read all battery parameters via serial communication
In the sript BatteryFunctions.ino single functions were created in
order to read each single battery parameter using the main function
ReadBatparam.

In the script Serialfunctions.ino it was created a new generic function
get_bat_param that it is called in the process_message function
depending on the function code.

In the UGV.ino script the battery parmaeters are read each one in
each iteration of the main loop to space operations over time and
reduce the cycle time.

Furthermore, some types of data were changed for keeping coherency.